Add files via upload
This work develops a convolutional
neural network model to perform binary
classification on images. To necessitate this,
first data preprocessing is performed wherein
the images are reduced to 64x64 images. A
CNN model is developed and this model is fit
to the training data. Based on the
performance of the model, optimization is
performed by tuning the model
hyperparameters. The model is run on the
training data for specified number of
EPOCHS till the best accuracy is reached.
The performance of the model is analyzed
using performance metrics like Accuracy. An
accuracy of 97% was obtained after 20
EPOCHS.